feat(pack): 可配置打包前是否清空输出目录

+++ src/models/Project.js
@@ -232,9 +232,11 @@ class Project {
                 config.devtool = opt.sourcemap
             }
 
-            try {
-                childProcess.execSync('rm -rf ' + config.output.path);
-            } catch (e) {}
+            if(opt.clean) {
+                try {
+                    childProcess.execSync('rm -rf ' + config.output.path);
+                } catch (e) {}
+            }
 
             config.plugins.push(new ProgressBarPlugin())
